+++ resources/views/livewire/table.blade.php
@@ -118,7 +118,10 @@
             </div>
         </div>
 
-        <x-model-browser::csv-buttons :$exportLimit />
+        <div class="d-flex justify-content-center align-items-start flex-wrap mt-3 gap-3">
+            <x-model-browser::csv-buttons :$exportLimit />
+            <x-model-browser::copy-page-button />
+        </div>
 
     </div>
 </div>